+++ januscloud/proxy/plugin/videoroom.py
@@ -706,7 +706,7 @@ class VideoRoomPublisher(object):
 
             # 1. leave the room
             try:
-                _send_backend_message(backend_handle, {
+                backend_handle.send_message({
                     'request': 'leave',
                 })
             except Exception:
@@ -1346,11 +1346,13 @@ class BackendRoom(object):
             # 1 destroy the room of the backend server
             if backend_handle:
                 # 2. destroy backend room
-                _send_backend_message(backend_handle, {
+                backend_handle.send_message({
                     'request': 'destroy',
                     'room': self.backend_room_id
                 })
-        except Exception:
+        except Exception as e:
+            log.exception('Backend room "{}"({}) failed to destroyed: {}, ignore'.
+                    format(self.backend_room_id, self.server_url, str(e)))             
             pass   # ignore destroy failed
         finally:
             # 3. detach handle
@@ -2207,100 +2209,6 @@ class VideoRoomManager(object):
                 # session timeout check is disable, just None loop
                 gevent.sleep(ROOM_CLEANUP_CHECK_INTERVAL)
 
-
-class VideoRoomBackendSweeper(object):
-    LIST_CHECK_INTERVAL = 60 # 10
-
-    def __init__(self, backend_server_mgr, idle_room_check_interval=600):
-        self._backend_server_mgr = backend_server_mgr
-        self._idle_room_check_interval = idle_room_check_interval
-        self._servers = {}
-        self._idle_room_check_greenlets = {}
-        self._server_list_check_greenlet = gevent.spawn(self._server_list_check_routine)
-        self._has_destroy = False
-
-    def destroy(self):
-        if self._has_destroy:
-            return
-        self._has_destroy = True
-        self._server_list_check_greenlet = None
-        self._idle_room_check_greenlets.clear()
-        self._servers.clear()
-
-    def _server_list_check_routine(self):
-        while not self._has_destroy:
-            gevent.sleep(VideoRoomBackendSweeper.LIST_CHECK_INTERVAL)
-            if self._has_destroy:
-                break
-
-            valid_server_list = self._backend_server_mgr.get_valid_server_list()
-            valid_server_names = {server.name for server in valid_server_list}
-            cur_server_names = set(self._servers)
-
-            # remove the invalid servers
-            server_names_to_remove = cur_server_names.difference(valid_server_names)
-            for name in server_names_to_remove:
-                self._servers.pop(name, None)
-                self._idle_room_check_greenlets.pop(name, None)
-
-            # add the new valid servers
-            for server in valid_server_list:
-                if server.name not in self._servers:
-                    self._servers[server.name] = server
-                    self._idle_room_check_greenlets[server.name] = gevent.spawn(
-                        self._idle_room_check_routine, server
-                    )
-
-    def _idle_room_check_routine(self, server):
-        idle_rooms = set()
-        backend_handle = None
-        while gevent.getcurrent() == self._idle_room_check_greenlets.get(server.name):
-            gevent.sleep(self._idle_room_check_interval)
-            if gevent.getcurrent() != self._idle_room_check_greenlets.get(server.name):
-                break
-            try:
-                # create backend handle
-                backend_handle = get_backend_session(
-                    server.url,
-                    auto_destroy=self._idle_room_check_interval * 2
-                ).attach_handle(JANUS_VIDEOROOM_PACKAGE)
-
-                reply_data, reply_jsep = _send_backend_message(backend_handle, {
-                    'request': 'list'
-                })
-                room_list_info = reply_data.get('list', [])
-                last_idle_rooms = idle_rooms
-                idle_rooms = set()
-                for room_info in room_list_info:
-                    room_id = room_info.get('room', 0)
-
-                    if room_id == 0 or room_info.get('num_participants', 1) > 0:
-                        continue   # pass invalid or in-service room
-                    if not room_info.get('description', '').startswith('januscloud'):
-                        continue   # pass not januscloud-created room
-
-                    # this is a idle room
-                    if room_id in last_idle_rooms:
-                        # auto destroy the idle room
-                        log.warning('Sweeper found the backend idle room {} for server {}, destroy it'.format(
-                            room_id, server.url))
-                        backend_handle.send_message({
-                            'request': 'destroy',
-                            'room': room_id
-                        })
-                    else:
-                        idle_rooms.add(room_id)  # destroy next time
-
-            except Exception as e:
-                # pass the exception
-                log.debug('Videoroom backend sweeper failed on server {}: {}. Retry in {} secs'.format(
-                    server.url, str(e), self._idle_room_check_interval))
-            finally:
-                if backend_handle:
-                    backend_handle.detach()
-                backend_handle = None
-
-
 class VideoRoomHandle(FrontendHandleBase):
 
     def __init__(self, handle_id, session, plugin, opaque_id=None, *args, **kwargs):
@@ -3051,13 +2959,6 @@ class VideoRoomPlugin(PluginBase):
         includeme(pyramid_config)
         pyramid_config.registry.videoroom_plugin = self
 
-        self.backend_sweeper = None
-        if self.config['backend_sweeper']['enable']:
-            self.backend_sweeper = VideoRoomBackendSweeper(
-                backend_server_mgr=backend_server_mgr,
-                idle_room_check_interval=self.config['backend_sweeper']['backend_idle_room_check_interval']
-            )
-
         log.info('{} initialized!'.format(JANUS_VIDEOROOM_NAME))
 
     def get_version(self):
@@ -3090,11 +2991,6 @@ class VideoRoomPlugin(PluginBase):
                 Optional("room_auto_destroy_timeout"): Default(IntVal(min=0, max=86400), default=0),
                 Optional("admin_key"): Default(StrVal(), default=''),
                 Optional("lock_rtp_forward"): Default(BoolVal(), default=False),
-                AutoDel(str): object  # for all other key we don't care
-            }, default={}),
-            Optional("backend_sweeper"): Default({
-                Optional("backend_idle_room_check_interval"): Default(IntVal(min=1, max=86400), default=600),
-                Optional("enable"): Default(BoolVal(), default=True),
                 AutoDel(str): object  # for all other key we don't care
             }, default={}),
             Optional("rooms"): Default([{