+++ januscloud/sentinel/main.py
@@ -38,6 +38,7 @@ def do_main(config):
     from januscloud.common.logger import set_root_logger
     from januscloud.sentinel.process_mngr import ProcWatcher
     from januscloud.sentinel.janus_server import JanusServer
+    from januscloud.sentinel.videoroom_sweeper import VideoroomSweeper
     from januscloud.sentinel.poster_manager import add_poster, list_posters
     from januscloud.sentinel.poster.http_poster import HttpPoster
 
@@ -47,6 +48,7 @@ def do_main(config):
     log = logging.getLogger(__name__)
 
     janus_watcher = None
+    videoroom_sweeper = None
     try:
         # set up janus server
         janus_server = JanusServer(
@@ -64,12 +66,24 @@ def do_main(config):
             isp=config['janus']['isp'],
         )
 
+        # set up videoroom_sweeper
+        if config['videoroom_sweeper']['enable']:
+            videoroom_sweeper = VideoroomSweeper(
+                server_ip=config['janus']['server_ip'],
+                ws_port=config['janus']['ws_port'],
+                des_filter=config['videoroom_sweeper']['des_filter'],
+                check_interval=config['videoroom_sweeper']['check_interval'],
+                room_auto_destroy_timeout=config['videoroom_sweeper']['room_auto_destroy_timeout'],
+            )
+            janus_server.register_listener(videoroom_sweeper)
+
         # set up janus_watcher
         if config['proc_watcher']['cmdline']:
             janus_watcher = ProcWatcher(args=shlex.split(config['proc_watcher']['cmdline']),
                                         error_restart_interval=config['proc_watcher']['error_restart_interval'],
                                         poll_interval=config['proc_watcher']['poll_interval'],
                                         process_status_cb=janus_server.on_process_status_change)
+            
 
         for poster_params in config['posters']:
             add_poster(janus_server, **poster_params)
@@ -93,6 +107,8 @@ def do_main(config):
         if janus_watcher:
             janus_watcher.start()
 
+        if videoroom_sweeper:
+            videoroom_sweeper.start()
 
         log.info('Janus Sentinel launched successfully')
 
@@ -101,6 +117,7 @@ def do_main(config):
             global _terminated
             _terminated = True
             rest_server.stop()
+
 
         gevent.signal_handler(signal.SIGTERM, stop_sentinel)
         gevent.signal_handler(signal.SIGQUIT, stop_sentinel)
@@ -111,6 +128,10 @@ def do_main(config):
         # while not _terminated:
         #    gevent.sleep(1)
 
+        if videoroom_sweeper:
+            videoroom_sweeper.destroy()
+            videoroom_sweeper = None
+
         # destroy janus server
         janus_server.destroy()
 
@@ -119,6 +140,7 @@ def do_main(config):
             janus_watcher.destroy()
             janus_watcher = None
 
+        
         # post to janux-proxy
         posters = list_posters()
         for poster in posters: